OPT-987 move browser projection cache contracts
